requirements.txt
# Auto Trading App Dependencies
# Install with: pip install -r requirements.txt

# Webull API (unofficial)
webull>=0.3.0

# Market Data
yfinance>=0.2.30,<1.0

# Technical Analysis
pandas>=2.0.0
numpy>=1.24.0
ta>=0.10.2  # Technical Analysis library

# Database
sqlalchemy>=2.0.0

# Scheduler
apscheduler>=3.10.0

# Notifications
twilio>=8.10.0

# Secure Credential Storage (macOS Keychain)
keyring>=24.0.0

# HTTP requests
requests>=2.31.0

# Configuration
pyyaml>=6.0

# Timezone handling
pytz>=2023.3

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
python-dotenv>=1.0.0  # For local dev environment (optional)
